+++ comments.php
@@ -1,6 +1,6 @@
 <?php
 /*
-The comments page for Bones
+The comments page for jenesis
 */
 
 // don't load it if you can't comment
@@ -14,7 +14,7 @@ if ( post_password_required() ) {
 
   <?php if ( have_comments() ) : ?>
 
-    <h3 id="comments-title" class="h2"><?php comments_number( __( '<span>No</span> Comments', 'bonestheme' ), __( '<span>One</span> Comment', 'bonestheme' ), _n( '<span>%</span> Comments', '<span>%</span> Comments', get_comments_number(), 'bonestheme' ) );?></h3>
+    <h3 id="comments-title" class="h2"><?php comments_number( __( '<span>No</span> Comments', 'jenesistheme' ), __( '<span>One</span> Comment', 'jenesistheme' ), _n( '<span>%</span> Comments', '<span>%</span> Comments', get_comments_number(), 'jenesistheme' ) );?></h3>
 
     <section class="commentlist">
       <?php
@@ -22,7 +22,7 @@ if ( post_password_required() ) {
           'style'             => 'div',
           'short_ping'        => true,
           'avatar_size'       => 40,
-          'callback'          => 'bones_comments',
+          'callback'          => 'jenesis_comments',
           'type'              => 'all',
           'reply_text'        => 'Reply',
           'page'              => '',
@@ -35,13 +35,13 @@ if ( post_password_required() ) {
 
     <?php if ( get_comment_pages_count() > 1 && get_option( 'page_comments' ) ) : ?>
     	<nav class="navigation comment-navigation" role="navigation">
-      	<div class="comment-nav-prev"><?php previous_comments_link( __( '&larr; Previous Comments', 'bonestheme' ) ); ?></div>
-      	<div class="comment-nav-next"><?php next_comments_link( __( 'More Comments &rarr;', 'bonestheme' ) ); ?></div>
+      	<div class="comment-nav-prev"><?php previous_comments_link( __( '&larr; Previous Comments', 'jenesistheme' ) ); ?></div>
+      	<div class="comment-nav-next"><?php next_comments_link( __( 'More Comments &rarr;', 'jenesistheme' ) ); ?></div>
     	</nav>
     <?php endif; ?>
 
     <?php if ( ! comments_open() ) : ?>
-    	<p class="no-comments"><?php _e( 'Comments are closed.' , 'bonestheme' ); ?></p>
+    	<p class="no-comments"><?php _e( 'Comments are closed.' , 'jenesistheme' ); ?></p>
     <?php endif; ?>
 
   <?php endif; ?>