+++ functions.php
@@ -1,64 +1,64 @@
 <?php
 /*
 Author: Eddie Machado
-URL: htp://themble.com/bones/
+URL: htp://themble.com/jenesis/
 
 This is where you can drop your custom functions or
 just edit things like thumbnail sizes, header images,
 sidebars, comments, ect.
 */
 
-// LOAD BONES CORE (if you remove this, the theme will break)
-require_once( 'library/bones.php' );
+// LOAD jenesis CORE (if you remove this, the theme will break)
+require_once( 'library/jenesis.php' );
 
 // USE THIS TEMPLATE TO CREATE CUSTOM POST TYPES EASILY
 require_once( 'library/custom-post-type.php' );
 
 // CUSTOMIZE THE WORDPRESS ADMIN (off by default)
-require_once( 'library/admin.php' );
+// require_once( 'library/admin.php' );
 
 /*********************
-LAUNCH BONES
+LAUNCH jenesis
 Let's get everything up and running.
 *********************/
 
-function bones_ahoy() {
+function jenesis_ahoy() {
 
   // let's get language support going, if you need it
-  load_theme_textdomain( 'jenesis', get_template_directory() . '/library/translation' );
+  load_theme_textdomain( 'jenesistheme', get_template_directory() . '/library/translation' );
 
   // launching operation cleanup
-  add_action( 'init', 'bones_head_cleanup' );
+  add_action( 'init', 'jenesis_head_cleanup' );
   // A better title
   add_filter( 'wp_title', 'rw_title', 10, 3 );
   // remove WP version from RSS
-  add_filter( 'the_generator', 'bones_rss_version' );
+  add_filter( 'the_generator', 'jenesis_rss_version' );
   // remove pesky injected css for recent comments widget
-  add_filter( 'wp_head', 'bones_remove_wp_widget_recent_comments_style', 1 );
+  add_filter( 'wp_head', 'jenesis_remove_wp_widget_recent_comments_style', 1 );
   // clean up comment styles in the head
-  add_action( 'wp_head', 'bones_remove_recent_comments_style', 1 );
+  add_action( 'wp_head', 'jenesis_remove_recent_comments_style', 1 );
   // clean up gallery output in wp
-  add_filter( 'gallery_style', 'bones_gallery_style' );
+  add_filter( 'gallery_style', 'jenesis_gallery_style' );
 
   // enqueue base scripts and styles
-  add_action( 'wp_enqueue_scripts', 'bones_scripts_and_styles', 999 );
+  add_action( 'wp_enqueue_scripts', 'jenesis_scripts_and_styles', 999 );
   // ie conditional wrapper
 
   // launching this stuff after theme setup
-  bones_theme_support();
+  jenesis_theme_support();
 
   // adding sidebars to Wordpress (these are created in functions.php)
-  add_action( 'widgets_init', 'bones_register_sidebars' );
+  add_action( 'widgets_init', 'jenesis_register_sidebars' );
 
   // cleaning up random code around images
-  add_filter( 'the_content', 'bones_filter_ptags_on_images' );
+  add_filter( 'the_content', 'jenesis_filter_ptags_on_images' );
   // cleaning up excerpt
-  add_filter( 'excerpt_more', 'bones_excerpt_more' );
+  add_filter( 'excerpt_more', 'jenesis_excerpt_more' );
 
-} /* end bones ahoy */
+} /* end jenesis ahoy */
 
 // let's get this party started
-add_action( 'after_setup_theme', 'bones_ahoy' );
+add_action( 'after_setup_theme', 'jenesis_ahoy' );
 
 
 /************* OEMBED SIZE OPTIONS *************/
@@ -70,8 +70,8 @@ if ( ! isset( $content_width ) ) {
 /************* THUMBNAIL SIZE OPTIONS *************/
 
 // Thumbnail sizes
-add_image_size( 'bones-thumb-600', 600, 150, true );
-add_image_size( 'bones-thumb-300', 300, 100, true );
+add_image_size( 'jenesis-thumb-600', 600, 150, true );
+add_image_size( 'jenesis-thumb-300', 300, 100, true );
 
 /*
 to add more sizes, simply copy a line from above
@@ -85,20 +85,20 @@ inside the thumbnail function.
 
 For example, to call the 300 x 300 sized image,
 we would use the function:
-<?php the_post_thumbnail( 'bones-thumb-300' ); ?>
+<?php the_post_thumbnail( 'jenesis-thumb-300' ); ?>
 for the 600 x 100 image:
-<?php the_post_thumbnail( 'bones-thumb-600' ); ?>
+<?php the_post_thumbnail( 'jenesis-thumb-600' ); ?>
 
 You can change the names and dimensions to whatever
 you like. Enjoy!
 */
 
-add_filter( 'image_size_names_choose', 'bones_custom_image_sizes' );
+add_filter( 'image_size_names_choose', 'jenesis_custom_image_sizes' );
 
-function bones_custom_image_sizes( $sizes ) {
+function jenesis_custom_image_sizes( $sizes ) {
     return array_merge( $sizes, array(
-        'bones-thumb-600' => __('600px by 150px'),
-        'bones-thumb-300' => __('300px by 100px'),
+        'jenesis-thumb-600' => __('600px by 150px'),
+        'jenesis-thumb-300' => __('300px by 100px'),
     ) );
 }
 
@@ -113,11 +113,11 @@ new image size.
 /************* ACTIVE SIDEBARS ********************/
 
 // Sidebars & Widgetizes Areas
-function bones_register_sidebars() {
+function jenesis_register_sidebars() {
 	register_sidebar(array(
 		'id' => 'sidebar1',
-		'name' => __( 'Sidebar 1', 'bonestheme' ),
-		'description' => __( 'The first (primary) sidebar.', 'bonestheme' ),
+		'name' => __( 'Sidebar 1', 'jenesistheme' ),
+		'description' => __( 'The first (primary) sidebar.', 'jenesistheme' ),
 		'before_widget' => '<div id="%1$s" class="widget %2$s">',
 		'after_widget' => '</div>',
 		'before_title' => '<h4 class="widgettitle">',
@@ -134,8 +134,8 @@ function bones_register_sidebars() {
 
 	register_sidebar(array(
 		'id' => 'sidebar2',
-		'name' => __( 'Sidebar 2', 'bonestheme' ),
-		'description' => __( 'The second (secondary) sidebar.', 'bonestheme' ),
+		'name' => __( 'Sidebar 2', 'jenesistheme' ),
+		'description' => __( 'The second (secondary) sidebar.', 'jenesistheme' ),
 		'before_widget' => '<div id="%1$s" class="widget %2$s">',
 		'after_widget' => '</div>',
 		'before_title' => '<h4 class="widgettitle">',
@@ -154,7 +154,7 @@ function bones_register_sidebars() {
 /************* COMMENT LAYOUT *********************/
 
 // Comment Layout
-function bones_comments( $comment, $args, $depth ) {
+function jenesis_comments( $comment, $args, $depth ) {
    $GLOBALS['comment'] = $comment; ?>
   <div id="comment-<?php comment_ID(); ?>" <?php comment_class('cf'); ?>>
     <article  class="cf">
@@ -172,13 +172,13 @@ function bones_comments( $comment, $args, $depth ) {
         ?>
         <img data-gravatar="http://www.gravatar.com/avatar/<?php echo md5( $bgauthemail ); ?>?s=40" class="load-gravatar avatar avatar-48 photo" height="40" width="40" src="<?php echo get_template_directory_uri(); ?>/library/images/nothing.gif" />
         <?php // end custom gravatar call ?>
-        <?php printf(__( '<cite class="fn">%1$s</cite> %2$s', 'bonestheme' ), get_comment_author_link(), edit_comment_link(__( '(Edit)', 'bonestheme' ),'  ','') ) ?>
-        <time datetime="<?php echo comment_time('Y-m-j'); ?>"><a href="<?php echo htmlspecialchars( get_comment_link( $comment->comment_ID ) ) ?>"><?php comment_time(__( 'F jS, Y', 'bonestheme' )); ?> </a></time>
+        <?php printf(__( '<cite class="fn">%1$s</cite> %2$s', 'jenesistheme' ), get_comment_author_link(), edit_comment_link(__( '(Edit)', 'jenesistheme' ),'  ','') ) ?>
+        <time datetime="<?php echo comment_time('Y-m-j'); ?>"><a href="<?php echo htmlspecialchars( get_comment_link( $comment->comment_ID ) ) ?>"><?php comment_time(__( 'F jS, Y', 'jenesistheme' )); ?> </a></time>
 
       </header>
       <?php if ($comment->comment_approved == '0') : ?>
         <div class="alert alert-info">
-          <p><?php _e( 'Your comment is awaiting moderation.', 'bonestheme' ) ?></p>
+          <p><?php _e( 'Your comment is awaiting moderation.', 'jenesistheme' ) ?></p>
         </div>
       <?php endif; ?>
       <section class="comment_content cf">
@@ -198,12 +198,12 @@ external fonts. If you're using Google Fonts, you
 can replace these fonts, change it in your scss files
 and be up and running in seconds.
 */
-function bones_fonts() {
+function jenesis_fonts() {
   wp_register_style('googleFonts', 'http://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic');
   wp_enqueue_style( 'googleFonts');
 }
 
-add_action('wp_print_styles', 'bones_fonts');
+add_action('wp_print_styles', 'jenesis_fonts');