+++ library/admin.php
@@ -7,7 +7,7 @@ It's turned off by default, but you can call it
 via the functions file.
 
 Developed by: Eddie Machado
-URL: http://themble.com/bones/
+URL: http://themble.com/jenesis/
 
 Special Thanks for code & inspiration to:
 @jackmcconnell - http://www.voltronik.co.uk/
@@ -43,7 +43,7 @@ function disable_default_dashboard_widgets() {
 	have more plugin widgets you'd like to remove?
 	share them with us so we can get a list of
 	the most commonly used. :D
-	https://github.com/eddiemachado/bones/issues
+	https://github.com/eddiemachado/jenesis/issues
 	*/
 }
 
@@ -59,7 +59,7 @@ http://digwp.com/2010/10/customize-wordpress-dashboard/
 */
 
 // RSS Dashboard Widget
-function bones_rss_dashboard_widget() {
+function jenesis_rss_dashboard_widget() {
 	if ( function_exists( 'fetch_feed' ) ) {
 		// include_once( ABSPATH . WPINC . '/feed.php' );               // include the required file
 		$feed = fetch_feed( 'http://themble.com/feed/rss/' );        // specify the source feed
@@ -70,7 +70,7 @@ function bones_rss_dashboard_widget() {
 	else foreach ($items as $item) { ?>
 
 	<h4 style="margin-bottom: 0;">
-		<a href="<?php echo $item->get_permalink(); ?>" title="<?php echo mysql2date( __( 'j F Y @ g:i a', 'bonestheme' ), $item->get_date( 'Y-m-d H:i:s' ) ); ?>" target="_blank">
+		<a href="<?php echo $item->get_permalink(); ?>" title="<?php echo mysql2date( __( 'j F Y @ g:i a', 'jenesistheme' ), $item->get_date( 'Y-m-d H:i:s' ) ); ?>" target="_blank">
 			<?php echo $item->get_title(); ?>
 		</a>
 	</h4>
@@ -81,8 +81,8 @@ function bones_rss_dashboard_widget() {
 }
 
 // calling all custom dashboard widgets
-function bones_custom_dashboard_widgets() {
-	wp_add_dashboard_widget( 'bones_rss_dashboard_widget', __( 'Recently on Themble (Customize on admin.php)', 'bonestheme' ), 'bones_rss_dashboard_widget' );
+function jenesis_custom_dashboard_widgets() {
+	wp_add_dashboard_widget( 'jenesis_rss_dashboard_widget', __( 'Recently on Themble (Customize on admin.php)', 'jenesistheme' ), 'jenesis_rss_dashboard_widget' );
 	/*
 	Be sure to drop any other created Dashboard Widgets
 	in this function and they will all load.
@@ -93,7 +93,7 @@ function bones_custom_dashboard_widgets() {
 // removing the dashboard widgets
 add_action( 'admin_menu', 'disable_default_dashboard_widgets' );
 // adding any custom widgets
-add_action( 'wp_dashboard_setup', 'bones_custom_dashboard_widgets' );
+add_action( 'wp_dashboard_setup', 'jenesis_custom_dashboard_widgets' );
 
 
 /************* CUSTOM LOGIN PAGE *****************/
@@ -102,20 +102,20 @@ add_action( 'wp_dashboard_setup', 'bones_custom_dashboard_widgets' );
 
 //Updated to proper 'enqueue' method
 //http://codex.wordpress.org/Plugin_API/Action_Reference/login_enqueue_scripts
-function bones_login_css() {
-	wp_enqueue_style( 'bones_login_css', get_template_directory_uri() . '/library/css/login.css', false );
+function jenesis_login_css() {
+	wp_enqueue_style( 'jenesis_login_css', get_template_directory_uri() . '/library/css/login.css', false );
 }
 
 // changing the logo link from wordpress.org to your site
-function bones_login_url() {  return home_url(); }
+function jenesis_login_url() {  return home_url(); }
 
 // changing the alt text on the logo to show your site name
-function bones_login_title() { return get_option( 'blogname' ); }
+function jenesis_login_title() { return get_option( 'blogname' ); }
 
 // calling it only on the login page
-add_action( 'login_enqueue_scripts', 'bones_login_css', 10 );
-add_filter( 'login_headerurl', 'bones_login_url' );
-add_filter( 'login_headertitle', 'bones_login_title' );
+add_action( 'login_enqueue_scripts', 'jenesis_login_css', 10 );
+add_filter( 'login_headerurl', 'jenesis_login_url' );
+add_filter( 'login_headertitle', 'jenesis_login_title' );
 
 
 /************* CUSTOMIZE ADMIN *******************/
@@ -128,11 +128,11 @@ you like.
 */
 
 // Custom Backend Footer
-function bones_custom_admin_footer() {
-	_e( '<span id="footer-thankyou">Developed by <a href="http://yoursite.com" target="_blank">Your Site Name</a></span>. Built using <a href="http://themble.com/bones" target="_blank">Bones</a>.', 'bonestheme' );
+function jenesis_custom_admin_footer() {
+	_e( '<span id="footer-thankyou">Developed by <a href="http://yoursite.com" target="_blank">Your Site Name</a></span>. Built using <a href="http://themble.com/jenesis" target="_blank">jenesis</a>.', 'jenesistheme' );
 }
 
 // adding it to the admin area
-add_filter( 'admin_footer_text', 'bones_custom_admin_footer' );
+add_filter( 'admin_footer_text', 'jenesis_custom_admin_footer' );
 
 ?>