+++ post-formats/format-status.php
@@ -7,7 +7,7 @@
                   <h1 class="entry-title single-title" itemprop="headline"><?php the_title(); ?></h1>
 
                   <p class="byline vcard">
-                    <?php printf( __( 'Posted <time class="updated" datetime="%1$s" pubdate>%2$s</time> by <span class="author">%3$s</span>', 'bonestheme' ), get_the_time('Y-m-j'), get_the_time(get_option('date_format')), get_the_author_link( get_the_author_meta( 'ID' ) )); ?>
+                    <?php printf( __( 'Posted <time class="updated" datetime="%1$s" pubdate>%2$s</time> by <span class="author">%3$s</span>', 'jenesistheme' ), get_the_time('Y-m-j'), get_the_time(get_option('date_format')), get_the_author_link( get_the_author_meta( 'ID' ) )); ?>
                   </p>
 
                 </header> <?php // end article header ?>
@@ -30,7 +30,7 @@
                      *
                     */
                     wp_link_pages( array(
-                      'before'      => '<div class="page-links"><span class="page-links-title">' . __( 'Pages:', 'bonestheme' ) . '</span>',
+                      'before'      => '<div class="page-links"><span class="page-links-title">' . __( 'Pages:', 'jenesistheme' ) . '</span>',
                       'after'       => '</div>',
                       'link_before' => '<span>',
                       'link_after'  => '</span>',
@@ -39,7 +39,7 @@
                 </section> <?php // end article section ?>
 
                 <footer class="article-footer">
-                  <?php the_tags( '<p class="tags"><span class="tags-title">' . __( 'Tags:', 'bonestheme' ) . '</span> ', ', ', '</p>' ); ?>
+                  <?php the_tags( '<p class="tags"><span class="tags-title">' . __( 'Tags:', 'jenesistheme' ) . '</span> ', ', ', '</p>' ); ?>
 
                 </footer> <?php // end article footer ?>